requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alembic==1.7.7
altair==4.2.2
appdirs==1.4.4
arrow==1.2.2
astroid==2.11.5
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
asyncssh==2.11.0
atomicwrites==1.4.0
atpublic==3.0.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.0
blinker==1.4
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy==0.7.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cookiecutter==2.1.1
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.28
databricks-cli==0.16.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.5.1
diskcache==5.4.0
distlib==0.3.4
distro==1.7.0
docker==4.2.2
dpath==2.0.6
dulwich==0.20.43
dvc==2.10.2
dvc-render==0.0.5
dvclive==0.8.0
entrypoints==0.4
ephem==4.1.3
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
Flask==2.1.1
flatbuffers==1.12
flatten-dict==0.4.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
flufl.lock==7.0
fonttools==4.31.1
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grandalf==0.6
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.1.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
kneed==0.7.0
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.7.1
libclang==14.0.1
mailchecker==4.1.17
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
mlflow==1.24.0
multidict==6.0.2
nanotime==0.5.2
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.4
networkx==2.8.4
nltk==3.8.1
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas @ file:///C:/b/abs_c6fuezktfm/croot/pandas_1670425103552/work
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
phonenumbers==8.12.50
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
pmdarima==1.8.5
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.8.0
pydot==1.4.2
PyDrive2==1.10.1
pyee==8.2.2
pygit2==1.9.2
Pygments==2.11.2
pygtrie==2.4.2
PyJWT==2.3.0
pylint==2.14.1
PyMeeus==0.5.11
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypiwin32==223
pyppeteer==1.0.2
pyrsistent==0.18.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
python-benedict==0.25.1
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn @ file:///C:/b/abs_61gq_04cwx/croot/scikit-learn_1673957229218/work
scipy==1.8.0
scmrepo==0.0.19
seaborn==0.12.0
semver==2.13.0
setuptools-git==1.2
shortuuid==1.0.9
shtab==1.5.4
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
split-folders==0.5.1
SQLAlchemy==1.4.34
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.17.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
toolz==0.12.0
tornado==6.1
tox==3.25.0
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.14.1
voluptuous==0.13.1
waitress==2.1.1
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.2
Werkzeug==2.1.1
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
wrapt==1.14.1
xgboost==1.5.2
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.0
